+++ Services/TicketStore.Web/Client/src/components/turnstile/status/Description.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import './Status.css';
+
+export const Description = ({ message }: { message: string }) => {
+  return (
+    <span id="status-description" className="description">{message}</span>
+  )
+};
